arrumando o calculo de parcelas

+++ src/main/java/Autofinanzi/dtos/VeiculoDto.java
@@ -2,6 +2,6 @@ package Autofinanzi.dtos;
 
 import Autofinanzi.models.enums.StatusVeiculo;
 
-public record VeiculoDto(String marca, String modelo, int valor, StatusVeiculo statusVeiculo) {
+public record VeiculoDto(String marca, String modelo, double valor, StatusVeiculo statusVeiculo) {
 
 }
